+++ packages/components/CHANGELOG.md
@@ -5,7 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [14.0.0] - 2026-03-02
+## [13.31.2] - 2026-03-04
+
+### Fixed
+
+- Added nullguards to (`data-table`, `dropdown`, `floor-selector`, `keyboard`, `list`, `my-position`, `route-instructions`, `scroll-buttons`, `search`, `share-sms`) components
+
+## [13.31.1] - 2026-03-03
 
 ### Changed